# server/http/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Technical Test KStyle Hub Backend API docs",
        "title": "Technical Test KStyle Hub API Backend",
        "contact": {
            "name": "Fawzi Linggo",
            "email": "fawzilinggo@google.com"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
...
        "/product": {
            "get": {
                "description": "Get All Product",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get All Product",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenters.GetProductResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit Product By ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Edit Product By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenters.EditProductByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Product",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Create Product",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenters.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Product By ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            }
        },
        "/product/reveiw": {
            "post": {
                "description": "Create Product Review",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ProductReview"
                ],
                "summary": "Create Product Review",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenters.CreateProductReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            }
        },
        "/product/review": {
            "get": {
                "description": "Get All ProductReview",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ProductReview"
                ],
                "summary": "Get All ProductReview",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenters.GetProductReviewResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit ProductReview By ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ProductReview"
                ],
                "summary": "Edit ProductReview By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenters.EditProductReviewByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete ProductReview By ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ProductReview"
                ],
                "summary": "Delete ProductReview By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            }
        },
        "/product/review/cancel-like": {
            "delete": {
                "description": "Delete LikeReview",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "LikeReview"
                ],
                "summary": "Delete LikeReview",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenters.CancelLikeReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            }
        },
        "/product/review/like": {
            "post": {
                "description": "Create LikeReview",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "LikeReview"
                ],
                "summary": "Create LikeReview",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenters.CreateLikeReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "description": "Get Product By ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get Product By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.GetProductByIDResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "presenters.BasicResponse": {
            "description": "General basic response",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/presenters.CreateEditResponse"
                },
                "error": {
                    "description": "Error will be empty if SUCCESS",
                    "type": "string",
                    "example": "ERROR or NULL"
                },
                "message": {
                    "type": "string",
                    "example": "FAILED or SUCCESS and etc"
                },
                "status": {
                    "description": "Status turn true if process SUCCESS",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "presenters.CancelLikeReviewRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "presenters.CreateEditResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "presenters.CreateLikeReviewRequest": {
            "type": "object",
            "required": [
                "member_id",
                "product_review_id"
            ],
            "properties": {
                "member_id": {
                    "type": "string"
                },
                "product_review_id": {
                    "type": "string"
                }
            }
        },
...
        "presenters.CreateProductRequest": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "presenters.CreateProductReviewRequest": {
            "type": "object",
            "required": [
                "desc_review",
                "member_id",
                "product_id"
            ],
            "properties": {
                "desc_review": {
                    "type": "string"
                },
                "member_id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
        "presenters.DataReview": {
            "type": "object",
            "properties": {
                "desc_review": {
                    "type": "string",
                    "example": "It's good"
                },
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "skin_color": {
                    "type": "string",
                    "example": "light"
                },
                "skin_type": {
                    "type": "string",
                    "example": "oily"
                },
                "total_like": {
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "fawzi"
                }
            }
        },
...
        "presenters.EditProductByIDRequest": {
            "type": "object",
            "required": [
                "id",
                "name",
                "price"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "presenters.EditProductReviewByIDRequest": {
            "type": "object",
            "required": [
                "desc_review",
                "id",
                "member_id",
                "product_id"
            ],
            "properties": {
                "desc_review": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "member_id": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
...
        "presenters.GetProductByIDResponse": {
            "type": "object",
            "properties": {
                "data_reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenters.DataReview"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "c836c745-e1db-4777-9699-e40fffaa302d"
                },
                "name": {
                    "type": "string",
                    "example": "Product 1"
                },
                "price": {
                    "type": "number",
                    "example": 100000
                },
                "total_like_review": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "presenters.GetProductResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "c836c745-e1db-4777-9699-e40fffaa302d"
                },
                "name": {
                    "type": "string",
                    "example": "Product 1"
                },
                "price": {
                    "type": "number",
                    "example": 100000
                }
            }
        },
        "presenters.GetProductReviewResponse": {
            "type": "object",
            "properties": {
                "desc_review": {
                    "type": "string",
                    "example": "Product 1"
                },
                "id": {
                    "type": "string",
                    "example": "c836c745-e1db-4777-9699-e40fffaa302d"
                },
                "member_id": {
                    "type": "string",
                    "example": "c836c745-e1db-4777-9699-e40fffaa302d"
                },
                "product_id": {
                    "type": "string",
                    "example": "c836c745-e1db-4777-9699-e40fffaa302d"
                }
            }
        }
    }
}